chore(ubergen): create missing L1 READMEs and strip stability banners
This change bundles together two changes for how ubergen handles package
READMEs. The first is to generate a base README for experimental/alpha modules
where only the L1s are included. The second is to strip out the stability
banners from *all* READMEs; this is because, by definition, everything included
in aws-cdk-lib is stable, so the banners are now redundant/misleading.

fixes #16567